# docs/contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: DataX Enterprise Studio API
  version: 1.2.0
  license:
    name: License pending repository-owner decision
    url: https://github.com/xiaoli2hust/datax/blob/main/README.md
  description: |
    V1 单组织、多项目、安全的一次性离线全量复制控制 API。
    仅支持 MySQL 8/PostgreSQL 15、人工触发、源端静默确认、预存在且实测为空的目标表和 insert-only-once。
    不提供调度、DAG、告警、AI、CDC、自由 SQL、Transformer 或插件安装接口。
    交付形态为 Windows 11 x64 本地工作站；业务 API 仅经 web 反向代理绑定本机
    127.0.0.1:17860，api、worker 和 postgres 不暴露宿主端口。
servers:
  - url: http://127.0.0.1:17860/api/v1
    description: Windows 本地工作站唯一支持入口；不得替换为 0.0.0.0、LAN IP 或公网地址
...
security:
  - bearerAuth: []
paths:
...
  /endpoint-policies/{endpoint_policy_id}:
    parameters:
      - name: endpoint_policy_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [SecurityPolicies]
      operationId: getEndpointPolicy
      summary: Admin 获取端点策略
      responses:
        '200':
          description: 端点策略
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [SecurityPolicies]
      operationId: updateEndpointPolicy
      summary: Admin 更新或停用端点策略
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndpointPolicyPatch'
      responses:
        '200':
          description: 已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointPolicy'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /endpoint-policies/{endpoint_policy_id}/revisions/{revision_id}:
    parameters:
      - name: endpoint_policy_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
      - name: revision_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [SecurityPolicies]
      operationId: getEndpointPolicyRevision
      summary: Admin 获取不可变端点策略修订
      description: 原始 FQDN/IP、CIDR、端口和出口策略属于 Admin-only 配置，不向项目只读角色暴露。
      responses:
        '200':
          description: 不可变 EndpointPolicyRevision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointPolicyRevision'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /endpoint-connection-evidence/{connection_evidence_id}:
    parameters:
      - name: connection_evidence_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [SecurityPolicies]
      operationId: getEndpointConnectionEvidence
      summary: Admin 获取不可变端点连接证据
      description: |
        返回一次 TEST/METADATA/PREFLIGHT/DATAX/ORACLE/RECOVERY_PROBE 连接的 DNS、
        selected/peer IP 与 egress 决策证据。该响应包含网络定位信息，仅 Admin 可读；
        项目成员响应只返回 connection_evidence_id。
      responses:
        '200':
          description: EndpointConnectionEvidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EndpointConnectionEvidence'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/datasources:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [Datasources]
      operationId: listDatasources
      summary: 分页列出项目数据源
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: engine
          in: query
          schema:
            $ref: '#/components/schemas/Engine'
      responses:
        '200':
          description: 数据源页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourcePage'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Datasources]
      operationId: createDatasource
      summary: 创建数据源并加密凭证
      description: 仅 Admin；host/port 必须命中 ACTIVE EndpointPolicy，运行时仍需 DNS 重解析和出口策略复检。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourceCreate'
            example:
              name: 订单库
              description: 只读订单源
              endpoint_policy_id: 33333333-3333-4333-8333-333333333333
              engine: MYSQL_8
              host: mysql.internal.example
              port: 3306
              database_name: sales
              default_schema: sales
              username: datax_reader
              password: <write-only-password>
              ssl_mode: VERIFY_FULL
      responses:
        '201':
          description: 已创建；响应不包含 password
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceAdminDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /datasources/{datasource_id}:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    get:
      tags: [Datasources]
      operationId: getDatasource
      summary: 获取项目可读的数据源脱敏摘要
      description: 对所有项目角色使用同一最小披露形状；不返回真实 host、port、database_name、default_schema 或 username。
      responses:
        '200':
          description: 数据源
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceRedactedSummary'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Datasources]
      operationId: updateDatasource
      summary: 更新数据源或轮换密码
      description: |
        password 缺失表示安全复用当前 ACTIVE secret；null 或空字符串无效。任一非秘密连接
        定位字段变化时，服务端必须先按当前 ACTIVE EndpointPolicyRevision 完成 DNS、精确
        出口、TLS/认证与物理实例身份探针，再新增不可变 DatasourceRevision 并原子切换
        current_revision_id；探针失败时整次 PATCH 失败且不能保存新修订或新 secret。
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourcePatch'
      responses:
        '200':
          description: 已更新
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceAdminDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
    delete:
      tags: [Datasources]
      operationId: deleteDatasource
      summary: 软删除数据源
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      responses:
        '204':
          description: 已软删除
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /datasources/{datasource_id}/admin-detail:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    get:
      tags: [Datasources]
      operationId: getDatasourceAdminDetail
      summary: Admin 获取数据源连接配置详情
      description: |
        仅 Admin 可读取真实 host、port、database_name、default_schema 和 username。
        响应仍不得包含密码、密文、nonce、DEK、完整 JDBC URL 或可直接执行的 DataX 配置。
      responses:
        '200':
          description: Admin-only 数据源详情
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceAdminDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasources/{datasource_id}/revisions/{revision_id}:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
      - name: revision_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    get:
      tags: [Datasources]
      operationId: getDatasourceRevision
      summary: Admin 获取不可变数据源修订原始详情
      description: 真实连接定位字段为 Admin-only；项目只读角色使用 DatasourceRedactedSummary。
      responses:
        '200':
          description: 数据源修订；永不包含密码、secret ID 或连接串
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceRevisionAdminDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasources/{datasource_id}/credential-secrets:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    get:
      tags: [Datasources]
      operationId: listDatasourceCredentialSecrets
      summary: Admin 列出凭据版本的非秘密生命周期元数据
      description: 不返回 secret ID、密文、nonce、Envelope 密文或 KEK 材料。
      responses:
        '200':
          description: 凭据版本摘要
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CredentialSecretSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasources/{datasource_id}/credential-secrets/{secret_version}/status:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
      - name: secret_version
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    post:
      tags: [Datasources]
      operationId: changeDatasourceCredentialSecretStatus
      summary: Admin 退役或紧急撤销凭据版本
      description: |
        ACTIVE 可转为 RETIRED、REVOKED 或 COMPROMISED；REVOKED/COMPROMISED
        必须 fail closed，并为引用该版本的非终态 Execution/RecoveryProbe 建立紧急终止事实。
        该接口不能把任何历史版本重新变为 ACTIVE；配置新密码应走数据源更新接口并创建新版本。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialSecretStatusChange'
      responses:
        '200':
          description: 更新后的非秘密生命周期摘要
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CredentialSecretSummary'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/ValidationError'
  /datasources/{datasource_id}/grants:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    get:
      tags: [SecurityPolicies]
      operationId: listDatasourceUsageGrants
      summary: Admin 分页列出数据源使用授权
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 授权页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceUsageGrantPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /datasources/{datasource_id}/grants/{member_id}:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
      - name: member_id
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/Uuid'
    put:
      tags: [SecurityPolicies]
      operationId: replaceDatasourceUsageGrants
      summary: Admin 原子替换成员的源端/目标端使用授权
      description: 空 usages 撤销全部授权；授权不能绕过项目成员关系、EndpointPolicy 或 TransferPolicy。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DatasourceUsageGrantReplace'
      responses:
        '200':
          description: 当前有效授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceUsageGrantPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /datasources/{datasource_id}/test:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    post:
      tags: [Datasources]
      operationId: testDatasource
      summary: 使用已保存凭证执行受限连接测试
      description: 仅 Admin；连接前必须复检 ACTIVE EndpointPolicy、DNS 解析结果和网络出口限制。
      responses:
        '200':
          description: 测试已完成；连接失败通过 status=FAILED 表达
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DatasourceTestResult'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /datasources/{datasource_id}/schema/tables:
    parameters:
      - $ref: '#/components/parameters/DatasourceId'
    get:
      tags: [Datasources]
      operationId: listDatasourceTables
      summary: 读取表和列元数据
      parameters:
        - name: schema_name
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - name: table_name
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 128
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 表元数据页
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableSchemaPage'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/DatasourceConnectionFailed'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: des_refresh
      description: V1 loopback HTTP 使用 HttpOnly、SameSite=Strict、Path=/api/v1/auth 且无 Domain；若未来改为受信 loopback HTTPS 必须增加 Secure。
  headers:
...
    ETag:
      description: 弱 ETag，值为 W/"row_version"
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
...
  parameters:
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema:
        type: string
        pattern: '^W/"[1-9][0-9]*"$'
  responses:
...
    Forbidden:
      description: 无权限
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: 资源不存在或对当前主体不可见
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: 版本、幂等、状态或唯一约束冲突
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    ValidationError:
      description: 请求或业务语义校验失败
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
  schemas:
    Uuid:
      type: string
      format: uuid
...
    EndpointPolicy:
      type: object
      additionalProperties: false
      required:
        - id
        - organization_id
        - name
        - current_revision_id
        - current_revision_no
        - current_revision
        - status
        - row_version
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        organization_id:
          $ref: '#/components/schemas/Uuid'
        name:
          type: string
          minLength: 1
          maxLength: 128
        current_revision_id:
          $ref: '#/components/schemas/Uuid'
        current_revision_no:
          type: integer
          minimum: 1
        current_revision:
          $ref: '#/components/schemas/EndpointPolicyRevision'
        status:
          type: string
          enum: [ACTIVE, DISABLED]
        row_version:
          type: integer
          minimum: 1
    EndpointPolicyRevision:
      type: object
      additionalProperties: false
      description: 不可变网络准入策略；任何字段变化都创建新 revision，历史引用不跟随 current pointer。
      required:
        - id
        - endpoint_policy_id
        - revision_no
        - engine
        - host_kind
        - host_value
        - allowed_cidrs
        - allowed_ports
        - tls_required
        - dns_ttl_ceiling_seconds
        - resolver_policy_version
        - egress_policy_version
        - policy_hash
        - created_by
        - created_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        endpoint_policy_id:
          $ref: '#/components/schemas/Uuid'
        revision_no:
          type: integer
          minimum: 1
        engine:
          $ref: '#/components/schemas/Engine'
        host_kind:
          type: string
          enum: [EXACT_FQDN, EXACT_IP]
        host_value:
          type: string
          minLength: 1
          maxLength: 253
        allowed_cidrs:
          type: array
          minItems: 1
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 3
            maxLength: 49
        allowed_ports:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items:
            type: integer
            minimum: 1
            maximum: 65535
        tls_required:
          type: boolean
        dns_ttl_ceiling_seconds:
          type: integer
          minimum: 1
          maximum: 3600
        resolver_policy_version:
          type: string
          minLength: 1
          maxLength: 64
          readOnly: true
        egress_policy_version:
          type: string
          minLength: 1
          maxLength: 64
          readOnly: true
        policy_hash:
          $ref: '#/components/schemas/Sha256'
        created_by:
          $ref: '#/components/schemas/Uuid'
        created_at:
          $ref: '#/components/schemas/Timestamp'
...
    EndpointConnectionEvidence:
      type: object
      additionalProperties: false
      description: >-
        每次 TEST/METADATA/PREFLIGHT/DATAX/ORACLE/RECOVERY_PROBE 阶段的不可变
        DNS 与出口证据。DATAX 由独立 Java/JDBC 进程执行，当前平台不能从该进程可靠
        回传 socket peer；此时 peer_observation_status 为 ENFORCED_NOT_OBSERVED、
        peer_ip 为 null，selected_ip 仅表示已由精确 IP 内核租约强制的唯一允许目的地址，
        不得解释为观测到的 peer。
      required:
        - id
        - operation_kind
        - datasource_revision_id
        - endpoint_policy_revision_id
        - resolver_policy_version
        - resolved_ips
        - selected_ip
        - peer_observation_status
        - peer_ip
        - egress_policy_version
        - egress_evidence_hash
        - decision
        - evidence_hash
        - observed_at
      properties:
        id:
          $ref: '#/components/schemas/Uuid'
        operation_kind:
          type: string
          enum: [TEST, METADATA, PREFLIGHT, DATAX, ORACLE, RECOVERY_PROBE]
        datasource_revision_id:
          $ref: '#/components/schemas/Uuid'
        endpoint_policy_revision_id:
          $ref: '#/components/schemas/Uuid'
        resolver_policy_version:
          type: string
          minLength: 1
          maxLength: 64
        resolved_ips:
          type: array
          minItems: 1
          maxItems: 64
          uniqueItems: true
          items:
            oneOf:
              - type: string
                format: ipv4
              - type: string
                format: ipv6
        selected_ip:
          type: string
          minLength: 2
          maxLength: 45
        peer_observation_status:
          type: string
          enum: [OBSERVED, ENFORCED_NOT_OBSERVED]
        peer_ip:
          oneOf:
            - type: string
              minLength: 2
              maxLength: 45
            - type: 'null'
        egress_policy_version:
          type: string
          minLength: 1
          maxLength: 64
        egress_evidence_hash:
          $ref: '#/components/schemas/Sha256'
        decision:
          type: string
          enum: [ALLOWED, DENIED]
        evidence_hash:
          $ref: '#/components/schemas/Sha256'
        observed_at:
          $ref: '#/components/schemas/Timestamp'
...
    EndpointPolicyPatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
        host_kind:
          type: string
          enum: [EXACT_FQDN, EXACT_IP]
        host_value:
          type: string
          minLength: 1
          maxLength: 253
        allowed_cidrs:
          type: array
          minItems: 1
          maxItems: 64
          uniqueItems: true
          items:
            type: string
            minLength: 3
            maxLength: 49
        allowed_ports:
          type: array
          minItems: 1
          maxItems: 16
          uniqueItems: true
          items:
            type: integer
            minimum: 1
            maximum: 65535
        tls_required:
          type: boolean
        dns_ttl_ceiling_seconds:
          type: integer
          minimum: 1
          maximum: 3600
        status:
          type: string
          enum: [ACTIVE, DISABLED]
...